refactor: update package structure for top rated
