+++ data/src/main/java/com/london/data/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.london.data.di
 
 import com.london.data.local.database.dao.search.GenreInterestDao
-import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.home.TopRatedLocal
+import com.london.data.local.model.home.popular.PopularSectionLocal
 import com.london.data.local.model.recent.search.RecentSearchLocal
 import com.london.data.local.model.recent.viewed.RecentViewedLocal
 import com.london.data.local.model.recent.watched.RecentWatchedMovieLocal
@@ -12,7 +12,7 @@ import com.london.data.local.model.search.SearchMoviesLocal
 import com.london.data.local.model.search.SearchTvShowLocal
 import com.london.data.local.preference.AuthPreferences
 import com.london.data.local.source.LocalDataSource
-import com.london.data.local.source.home.popular.HomeLocalDataSource
+import com.london.data.local.source.home.HomeLocalDataSource
 import com.london.data.local.source.home.upcoming.UpComingLocalDataSource
 import com.london.data.local.source.recent.RecentDataSource
 import com.london.data.local.source.recent.watched.RecentWatchedDataSource
