+++ data/src/main/java/com/london/data/repository/home/popular/PopularRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.london.data.repository.home.popular
 
 import com.london.data.local.model.home.popular.PopularSectionLocal
-import com.london.data.local.source.home.popular.HomeLocalDataSource
+import com.london.data.local.source.home.HomeLocalDataSource
 import com.london.data.mapper.popular.toMovieEntity
 import com.london.data.mapper.popular.toPopularMovieSectionLocal
 import com.london.data.mapper.popular.toPopularMovies
