+++ data/src/main/java/com/london/data/repository/toprated/TopRatedMovieRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.london.data.repository.toprated
 
 import com.london.data.local.model.home.TopRatedLocal
-import com.london.data.local.source.home.popular.HomeLocalDataSource
+import com.london.data.local.source.home.HomeLocalDataSource
 import com.london.data.mapper.toprated.toEntity
 import com.london.data.mapper.toprated.toLocal
 import com.london.data.mapper.toprated.toMovieEntity
